requirements.txt
slackclient
python-dotenv
flask
slackeventsapi
requests